add users front part

+++ app/Http/Controllers/Admin/PostController.php
@@ -13,7 +13,7 @@ use Illuminate\Support\Facades\Storage;
 class PostController extends Controller
 {
     /**
-     * Display a listing of the resource.
+     * Display a listing of the resources.
      *
      * @return Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
@@ -24,7 +24,7 @@ class PostController extends Controller
     }
 
     /**
-     * Show the form for creating a new resource.
+     * Show the form for creating a new resources.
      *
      * @return \Illuminate\Http\Response
      */
@@ -36,7 +36,7 @@ class PostController extends Controller
     }
 
     /**
-     * Store a newly created resource in storage.
+     * Store a newly created resources in storage.
      *
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
@@ -67,7 +67,7 @@ class PostController extends Controller
     }
 
     /**
-     * Display the specified resource.
+     * Display the specified resources.
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
@@ -78,7 +78,7 @@ class PostController extends Controller
 //    }
 
     /**
-     * Show the form for editing the specified resource.
+     * Show the form for editing the specified resources.
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
@@ -93,7 +93,7 @@ class PostController extends Controller
     }
 
     /**
-     * Update the specified resource in storage.
+     * Update the specified resources in storage.
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
@@ -127,7 +127,7 @@ class PostController extends Controller
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Remove the specified resources from storage.
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
@@ -136,7 +136,8 @@ class PostController extends Controller
     {
         $post = Post::find($id);
         $post->tags()->sync([]);
-        Storage::delete($post->thumbnail);
+        !is_null($post->thumbnail) && Storage::delete($post->thumbnail);
+//        Storage::delete($post->thumbnail);
         $post->delete();
         return redirect()->route('posts.index')->with('success', 'Статья удалена');
     }
